+++ BestPrice.java
@@ -3,13 +3,13 @@ import java.util.HashMap;
 import org.javatuples.*;
 
 public class BestPrice {
-	HashMap<String, Pair<String, Integer>> best = new HashMap<String, Pair<String, Integer>>();
+	HashMap<String, Pair<String, Double>> best = new HashMap<String, Pair<String, Double>>();
 	
-	public void addTriplet(String food, String location, Integer price) {
-		best.put(food, new Pair<String, Integer>(location, price));
+	public void addTriplet(String food, String location, Double price) {
+		best.put(food, new Pair<String, Double>(location, price));
 	}
 	
-	public Pair<String, Integer> getPrice(String food) {
+	public Pair<String, Double> getPrice(String food) {
 		return best.get(food);
 	}
