+++ PriceData.java
@@ -4,17 +4,17 @@ import java.util.HashMap;
 import org.javatuples.*;
 
 public class PriceData {
-	HashMap<String, HashMap<String, Integer>> database;
+	HashMap<String, HashMap<String, Double>> database;
 
-	public void addEntry(String food, String location, Integer price) {
+	public void addEntry(String food, String location, Double price) {
 		if (!database.containsKey(food)) {
-			database.put(food, new HashMap<String, Integer>(null));
+			database.put(food, new HashMap<String, Double>(null));
 		}
 		database.get(food).put(location, price);
 	}
 	
-	public void removeEntry(String food, String location, Integer price) {
-		Pair<String, Integer> toRemove = new Pair<String, Integer>(location, price);
+	public void removeEntry(String food, String location, Double price) {
+		Pair<String, Double> toRemove = new Pair<String, Double>(location, price);
 		database.get(food).remove(toRemove);
 	}
 	
@@ -22,18 +22,18 @@ public class PriceData {
 		for (String food : database.keySet()) {
 			for (String entry : database.get(food).keySet()){
 				String location = entry;
-				Integer price = database.get(food).get(entry);
+				Double price = database.get(food).get(entry);
 				System.out.println(food+" costs $"+price+" at "+location+".");
 			}
 				
 		}
 	}
 	
-	public int getPrice(String food, String location){
+	public Double getPrice(String food, String location){
 		return database.get(food).get(location);
 	}
 	
-	public HashMap<String, HashMap<String, Integer>> getDatabase() {
+	public HashMap<String, HashMap<String, Double>> getDatabase() {
 		return database;
 	}
